DEP24-P10-DeptDatabaseMgmtPortal-backend/requirements.txt
asgiref==3.5.2
dj-database-url==1.2.0
Django==3.2.18
django-cors-headers==3.13.0
django-environ==0.10.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
gunicorn==20.1.0
json2html==1.3.0
lxml==4.9.2
Markdown==3.4.1
numpy==1.26.4
openpyxl==3.0.10
pandas==2.2.0
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.6.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.9.0
tzdata==2022.7
whitenoise==6.2.0
